requirements.txt
streamlit
sentence-transformers
torch
spacy
pdfminer.six
docx2txt
numpy
scikit-learn
plotly
